Add Makefile and fix logging for failed -window assertion.
Get the program compiling and update README.md. The only
code change needed was to update a CHECK() call in main() to
log the original FLAGS_window string instead of the
istringstream that wraps it. Presumably operator<<() changed
in libglog such that it no longer accepts streams, or maybe
Chrome OS's libglog had a local patch to support streams.